Move User and CreateUser to package
